+++ CHANGELOG.md
@@ -82,6 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#136](https://github.com/thejoltjoker/kreacon/pull/136)
 - Add file validation to drop zone
   [#138](https://github.com/thejoltjoker/kreacon/pull/138)
+- Add admin ability to toggle user admin and role
+  [#149](https://github.com/thejoltjoker/kreacon/pull/149)
 
 ### Changed